+++ src/components/TripSpace/InviteFriends/InviteFriends.tsx
@@ -2,6 +2,8 @@ import { useState } from "react";
 
 import styles from "./InviteFriends.module.scss";
 
+import useKakaoShareButton from "@/hooks/useKakaoShareButton";
+
 import CopyClipboard from "@/components/Modal/CopyClipboard/CopyClipboard";
 
 import InviteKakao from "@/assets/inviteKakao.svg?react";
@@ -20,36 +22,6 @@ function InviteFriends() {
     }, 1000);
   };
 
-  const handleKakaoClick = () => {
-    window.Kakao.init(import.meta.env.VITE_KAKAO_KEY);
-    if (window.Kakao && window.Kakao.isInitialized()) {
-      window.Kakao.Share.createDefaultButton({
-        objectType: "feed",
-        container: "#kakaoShareButton",
-        content: {
-          title: "TRIPVOTE [강릉여행]",
-          description: "흰 돛과 일정만 있으면 어디든 갈 수 있어",
-          imageUrl:
-            "https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FNaMMW%2FbtsC8Vt9uj3%2FqKZw6iL4TNxA8YR4IFTmXk%2Fimg.png",
-          link: {
-            mobileWebUrl: "http://localhost:5173",
-            webUrl: "http://localhost:5173",
-          },
-        },
-
-        buttons: [
-          {
-            title: "같이 여행하기",
-            link: {
-              mobileWebUrl: "http://localhost:5173",
-              webUrl: "http://localhost:5173",
-            },
-          },
-        ],
-      });
-    }
-  };
-
   return (
     <>
       <div className={styles.container}>
@@ -64,7 +36,7 @@ function InviteFriends() {
           </p>
         </div>
         <div className={styles.container__wrapperButton}>
-          <button onClick={handleKakaoClick} id="kakaoShareButton">
+          <button onClick={useKakaoShareButton()} id="kakaoShareButton">
             <InviteKakao />
             <p className={styles.container__a11y}>카카오톡 초대</p>
           </button>
